+++ src/Symfony/Component/DependencyInjection/TypedReference.php
@@ -19,20 +19,24 @@ namespace Symfony\Component\DependencyInjection;
 class TypedReference extends Reference
 {
     private $type;
+    private $name;
     private $requiringClass;
 
     /**
      * @param string $id              The service identifier
      * @param string $type            The PHP type of the identified service
      * @param int    $invalidBehavior The behavior when the service does not exist
+     * @param string $name            The name of the argument targeting the service
      */
-    public function __construct(string $id, string $type, $invalidBehavior = ContainerInterface::EXCEPTION_ON_INVALID_REFERENCE)
+    public function __construct(string $id, string $type, $invalidBehavior = ContainerInterface::EXCEPTION_ON_INVALID_REFERENCE, $name = null)
     {
-        if (\is_string($invalidBehavior) || 3 < \func_num_args()) {
+        if (\is_string($invalidBehavior ?? '') || \is_int($name)) {
             @trigger_error(sprintf('The $requiringClass argument of "%s()" is deprecated since Symfony 4.1.', __METHOD__), E_USER_DEPRECATED);
 
             $this->requiringClass = $invalidBehavior;
             $invalidBehavior = 3 < \func_num_args() ? \func_get_arg(3) : ContainerInterface::EXCEPTION_ON_INVALID_REFERENCE;
+        } else {
+            $this->name = $type === $id ? $name : null;
         }
         parent::__construct($id, $invalidBehavior);
         $this->type = $type;
@@ -41,6 +45,11 @@ class TypedReference extends Reference
     public function getType()
     {
         return $this->type;
+    }
+
+    public function getName(): ?string
+    {
+        return $this->name;
     }
 
     /**